% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alimap.R
\docType{package}
\name{alimap}
\alias{alimap}
\alias{alimap-package}
\title{alimap: Chinese Maps from Aliyun}
\description{
A package to supply Chinese maps (sf objects)
from aliyun (http://datav.aliyun.com/tools/atlas).
}
\section{How to get the maps }{

map_nation() : get the national map of China.

map_province() : get the provincial map of China.

map_prefecture_city() : get the map of prefecture cities of China.
}

