% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefecture-city.R
\name{get_cities_map}
\alias{get_cities_map}
\title{get_cities_map}
\usage{
get_cities_map(adcode)
}
\arguments{
\item{adcode}{a numeric or a character of length one, the adcode of a province.}
}
\value{
a sf object, the boundary of prefecture cities for a province.
}
\description{
get_cities_map
}
