% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefecture-city.R
\name{get_city_map}
\alias{get_city_map}
\title{get_city_map}
\usage{
get_city_map(adcode)
}
\arguments{
\item{adcode}{a numeric or a character of length one, the adcode of a
province-level municipality or a special administrative region.}
}
\value{
a sf object, the boundary of a province-level municipality or
a special administrative region.
}
\description{
get_city_map
}
