% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nation.R
\name{map_nation}
\alias{map_nation}
\title{map_nation}
\usage{
map_nation()
}
\value{
a sf object of the national boundary of China.
}
\description{
get the national map of China.
}
\examples{
map <- map_nation()
plot(map)
}
