% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefecture-city.R
\name{map_prefecture_city}
\alias{map_prefecture_city}
\title{map_prefecture_city}
\usage{
map_prefecture_city()
}
\value{
a sf object, the boundary of all the prefecture cities of China.
}
\description{
get the map of prefecture cities of China.
}
\examples{
map <- map_prefecture_city()
plot(map)
}
