% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province.R
\name{map_province}
\alias{map_province}
\title{map_province}
\usage{
map_province()
}
\value{
a sf object of the provincial boundary of China.
}
\description{
get the provincial map of China.
}
\examples{
map <- map_province()
plot(map)
}
